+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## [version 1.4.0] 2024-03-14
 
+### Fixed
+
+- Demo data use negative amount for expenses
+
 ### Update
 
+- SanKey more usable on small devices (phones)
 - SanKey Net Lost now displays on the left side of Expenses
 - SanKey width is flush to the view port